Improvements: gitignore, Device Names as sources, Dry Run

+++ netbox-powerdns-sync.py
@@ -6,7 +6,8 @@ import re
 import ipaddress
 
 from config import NB_URL, NB_TOKEN, PDNS_API_URL, PDNS_KEY
-from config import FORWARD_ZONES, REVERSE_ZONES
+from config import FORWARD_ZONES, REVERSE_ZONES, DRY_RUN
+from config import SOURCE_DEVICE, SOURCE_IP
 
 nb = pynetbox.api(NB_URL, token=NB_TOKEN)
 
@@ -18,20 +19,36 @@ record_ips = []
 for forward_zone in FORWARD_ZONES:
     forward_zone_canonical = forward_zone+"."
 
-    # get IPs with DNS name ending in forward_zone from NetBox
-    nb_ips = nb.ipam.ip_addresses.filter(dns_name__iew=forward_zone)
+    # Source IP: Create domains based on domain attached directly to IPs
+    if SOURCE_IP:
+        # get IPs with DNS name ending in forward_zone from NetBox
+        nb_ips = nb.ipam.ip_addresses.filter(dns_name__iew=forward_zone)
 
-    # assemble list with tupels containing the canonical name, the record type
-    # and the IP address without the subnet from NetBox IPs
-    for nb_ip in nb_ips:
-        if nb_ip.family.value == 6:
-            type = "AAAA"
-        else:
-            type = "A"
-        host_ips.append((nb_ip.dns_name+".",
-                         type,
-                         re.sub("/[0-9]*", "", str(nb_ip)),
-                         forward_zone_canonical))
+        # assemble list with tupels containing the canonical name, the record type
+        # and the IP address without the subnet from NetBox IPs
+        for nb_ip in nb_ips:
+            if nb_ip.family.value == 6:
+                type = "AAAA"
+            else:
+                type = "A"
+            host_ips.append((nb_ip.dns_name+".",
+                            type,
+                            re.sub("/[0-9]*", "", str(nb_ip)),
+                            forward_zone_canonical))
+
+    # if enabled, iterate over all devices and set their primary IPs    
+    if SOURCE_DEVICE:
+        for d in nb.dcim.devices.filter(name__ic=forward_zone):
+            if d.primary_ip4:
+                host_ips.append((d.name+".",
+                    "A",
+                    re.sub("/[0-9]*", "", str(d.primary_ip4)),
+                    forward_zone_canonical))
+            if d.primary_ip6:
+                host_ips.append((d.name+".",
+                    "AAAA",
+                    re.sub("/[0-9]*", "", str(d.primary_ip6)),
+                    forward_zone_canonical))
 
     # get zone forward_zone_canonical form PowerDNS
     zone = pdns.get_zone(forward_zone_canonical)
@@ -102,26 +119,29 @@ for record in to_delete:
 
 print("----")
 
-for record in to_create:
-    print("Creating", record)
-    zone = pdns.get_zone(record[3])
-    zone.create_records([
-                        powerdns.RRSet(record[0],
-                                       record[1],
-                                       [(record[2], False)],
-                                       comments=[powerdns.Comment("NetBox")])
-                        ])
+if DRY_RUN == False:
+    for record in to_create:
+        print("Creating", record)
+        zone = pdns.get_zone(record[3])
+        zone.create_records([
+                            powerdns.RRSet(record[0],
+                                        record[1],
+                                        [(record[2], False)],
+                                        comments=[powerdns.Comment("NetBox")])
+                            ])
 
-print("----")
+    print("----")
 
-for record in to_delete:
-    print("Deleting", record)
-    zone = pdns.get_zone(record[3])
-    zone.delete_records([
-                        powerdns.RRSet(record[0],
-                                       record[1],
-                                       [(record[2], False)],
-                                       comments=[powerdns.Comment("NetBox")])
-                        ])
+    for record in to_delete:
+        print("Deleting", record)
+        zone = pdns.get_zone(record[3])
+        zone.delete_records([
+                            powerdns.RRSet(record[0],
+                                        record[1],
+                                        [(record[2], False)],
+                                        comments=[powerdns.Comment("NetBox")])
+                            ])
+else:
+    print("Skipping Create/Delete due to Dry Run")
 
 print("----")
